requirements.txt
streamlit
numpy
pillow
opencv-python-headless
scipy
scikit-image